server/requirements.txt
torch
torchvision
pillow==10.3.0
fastapi==0.115.4
uvicorn==0.32.0
python-multipart
aiofiles